Converted code to django
