+++ test_requirements.txt
@@ -1,0 +1,8 @@
+ddt==1.1.1
+coverage==4.0.3
+edx-lint==0.4.0
+mock==1.3.0
+pep8==1.6.2
+pytz
+codecov
+tox>=2.3.1,<3.0.0
