+++ tox.ini
@@ -1,0 +1,23 @@
+[tox]
+envlist = py{27}-django{18,19,110,111}, quality
+
+[testenv]
+setenv =
+    # This allows us to reference settings.py
+    PYTHONPATH = {toxinidir}
+
+deps =
+    django18: Django>=1.8,<1.9
+    django19: Django>=1.9,<1.10
+    django110: Django>=1.10,<1.11
+    django111: Django>=1.11,<2.0
+    -rtest_requirements.txt
+
+commands =
+    coverage run ./manage.py test --settings=edx_sga.settings
+    coverage report
+
+[testenv:quality]
+commands =
+    pep8 --config=.pep8 edx_sga
+    pylint --rcfile=pylintrc edx_sga
